Include db_oid when writing/referencing the binary on disk
FunctionOids are unique within the database but not across it.
It's possible for two PL/Rust functions to share the same fn_oid
across different databases leading to overwriting the same function,
or accessing it.

To prevent this include db_oid when constructing the name to
guarantee uniqueness of the files