ARM64 CI build (#77)
Adds the Github Actions setup for use with a custom self-hosted runner. The custom runner was required since Github does not offer aarch64 out of the box (at the time of this writing, that is).

Note that the underlying architecture to spin up and maintain the self-hosted runner are in another separate internal TCDI Github repository that is not available to the public.