Fix Rust highlighting for cargo expand
If cargo expand is used on plrust, it embeds the README.md
in the expansion inside a block doc-comment. A Rust syntax
highlighter will thus interpret the bash in the README.md
as Rust. This means some form of glob syntax are seen as
ending the comment instead.

The result is that everything has broken highlighting,
and if we ever wanted to actually include the README.md,
it might even break compilation!

Fortunately, Rust defines /* */ to have balanced nesting rules.
Thus, putting a Rust block-comment-open inside an HTML comment,
which is invisible when the README.markdown is viewed as HTML,
fixes this for Rust syntax handlers that don't also know Markdown.