Allow specifying an allow-listed of dependencies
Provide a GUC that will specify the file path
that contains an allow-listed set of dependencies.
If it is not set, treat it as allowing
an arbitrary list of dependencies. This is useful
to allow administrators of PL/Rust to limit the
scope of dependencies that are allowed in their
environment.