Map Postgres' `NUMERIC` type to pgx' (new) `AnyNumeric`.
We can't map to a specific `Numeric<P, S>` as Postgres doesn't store the "typmod" values in the `pg_proc` catalog, so by the time we create the function, those are lost to us.

I don't view this as a big deal, tho.  Users can use `AnyNumeric::rescale<P, S>().into()` to return a `AnyNumeric` that's been truncated/expanded to a particular precision and scale.  Alternatively, it can be handled at the SQL level, which I suspect is what most people do today with languages like plpgsql.