Map Postgres' `NUMERIC` type to pgx' (new) `AnyNumeric`. (#115)
We can't map to a specific `Numeric<P, S>` as Postgres doesn't store "typmod"
values in the `pg_proc` catalog, so by the time we create the function, those
are lost to us.

I don't view this as a big deal, tho.  Users can use
`AnyNumeric::rescale<P, S>().into()` to return an `AnyNumeric` that's been
truncated/expanded to a particular precision and scale.  Alternatively, it can
be handled at the SQL level, which I suspect is what most people do today with
languages like plpgsql.