Move the ultimate storage of a plrust compiled function from disk to an extension table named `plrust.plrust_proc`.
The compilation process now cleans up after itself (we no longer leave any function-specific artifacts on disk, other than the shared `${PLRUST.WORK_DIR}target/` directory), and the .so bytes are inserted into `plrust.plrust_proc`, tagged with the host's "target_triple", as the compiled pl/rust extension understands them.

This also enables plrust to properly work with `DROP FUNCTION`, `DROP SCHEMA`, and `ALTER FUNCTION` when a `LANGUAGE plrust` function is involved in one of those statements.  For the first two, we remove the corresponding entires in our `plrust.plrust_proc` table, for the latter we let everything pass except changes to the `STRICT` property.

The interception of those utility statements is done via a hook, which now means plrust needs to be configured as (at least) a `shared_preload_libraries` entry in `postgresql.conf`.

`CREATE OR REPLACE FUNCTION` is simply handled through the normal language handler API.

When we load a "function.so" for the first execution, it is streamed from value in `plrust.plrust_proc` to a temporary file in the configured `plrust.work_dir` directory, dlopen()'d and then immediately deleted from disk.

Loading a function necessitates that we have one already compiled for the current host "triple target".  If one doesn't exist currently plrust will raise an ERROR.  A future PR will address an automatic re-compilation scheme.

Also of note, plrust previously would compile the user function with `-Ctarget-cpu=native`.  This is no longer the case as it needs binaries to be compatible across same-target-triple-but-different-and-incompatible-cpu-capabilities.  Perhaps in the future this can be addressed in a more explicit manner.