Work on cross-compilation support
Add a GUC called `plrust.compilation_targets` that is a comma-separated list of other rust targets for which a plrust function should be compiled.  Update code to compile the plrust function for each target.

Also change from tracking file paths for the on-disk .so to its actual bytes as a `Vec<u8>`.  Also remove a bunch of unused getter functions from most of the states.