Work on cross-compilation support (#155)
Add a GUC called `plrust.compilation_targets` that is a comma-separated list of other rust targets for which a plrust function should be compiled.  Update code to compile the plrust function for each target.

Also change from tracking file paths for the on-disk .so to its actual bytes as a `Vec<u8>`.  Also remove a bunch of unused getter functions from most of the states.

## Other Changes

* rejigger how we detect the host target triple name because of postgrestd
* move `target.rs` up to the crate root and adjust code accordingly
* move the `CompilationTarget` struct over into `target.rs`, cleanup usages, and change `crate::target::tuple()` to return a `CompilationTarget` instead of a String.
* `target_postgrestd` shouldn't be a default feature as it makes plrust really awkward to build for platforms like macOS where it's not even supported.