Set a `$PATH` when there isn't one
Debian-based distro-managed Postgres installations (such as Ubuntu) seem to clear the `$PATH` environment variable when Postgres starts.  This means that plrust can't find the system commands such as `cargo` and `cc` that it needs to compile user functions (there are probably others).

If the `$PATH` is missing (or the empty string), we'll hardcode it to `~/.cargo/bin:/usr/bin` when executing `cargo`.  However, if the user has set the `plrust.PATH_override` GUC, then we will **always** use that when executing `cargo`.