Fix issue #200
This adds a `<'a>` lifetime to all user functions so that the few cases where arguments are borrowed (`&'a str`) and return types might borrow (`SetOfIterator<'a, ...>)` can do so.

Also blindly adds the `#[allow(unused_lifetime)]` annotation to all instances of user functions so that compilation warnings won't happen in the cases where all args/return don't borrow.