fix issue #204: `pg_restore` works
To properly handle restoration via `pg_restore` we need to at least try and parse the function "code" given to us as if it were a JSON blob that matches our format in `pg_catalog.pg_proc.prosrc`.  If it does match, we pluck out the "src" property.

If it doesn't match then it's probably just a regular CREATE OR REPLACE statement, so we use it as-is.

This logic is encapsulated inside the new `maybe_extract_source_from_json` function.