Rename `trusted-pgx` to `plrust-trusted-pgx`
This changeset renames `trusted-pgx` to `plrust-trusted-pgx` so that we can publish it to crates.io while being good citizens.  The intent is for it to be versioned independently of both `pgx` and `plrust`.  It is likely that it'll need to evolve in ways that aren't necessarily tied to specific pgx/plrust releases.

We also introduce a new GUC, `plrust.trusted_pgx_version` so plrust administrators can change the default version.  Which, the default version, is whatever `plrust-trusted-pgx` is set to in this workspace at the time the plrust binary was built.  This value is sussed out by spawning `cargo tree ...` during plrust's new `build.rs` execution.

The `plrust-trusted-pgx` version is now stored in the `pg_proc.prosrc` catalog entry in a new, non-optional, property named "trusted_pgx_version".  This value is not used right now, but it may be necessary in the future should we support re-compilation.

There's now a `./publish.sh` script in the root that only publishes plrust-trusted-pgx.