Make GitHub recognize the PostgreSQL license
This is a minor edit to the license file, to enable GitHub to recognize
the PostgreSQL license, and prominently display it on the project's
GitHub page. The effect of this license can be seen in the
[pg_license_test][] project, and the screenshot in that project's
README.

This fix also addresses a significant deficiency in the current license.
The last paragraph in the current license text does not include TCDI's
name. This is fixed by replacing all occurrences of company names with
the text 'COPYRIGHT HOLDER'. This way, if any new copyright holders
emerge in future, they are all collectively represented by the text
'COPYRIGHT HOLDER'.

The changes required to convince GitHub:
1. Prefix the Copyright year line with 'Copyright (c)'.

    Starting the line with 'Portions' fails GitHub's test.

2. Replace ccopyright holders' names with text 'COPYRIGHT HOLDER'

    The long company names confuses GitHub. Even trying to use the
plural form 'COPYRIGHT HOLDERS' makes GitHub not recognize the license.

To ensure that the license text has not materially changed, one can use
the `--color-words` Git option, and see that nothing significant has
changed in the license file. E.g. `git diff --color-words HEAD~1`.

[pg_license_test]: https://github.com/DrPostgres/pg_license_test