Don't require a loaded function to have "compile lints" applied.
Loading an existing, compiled function should not require that it was
previously compiled with the set of lints PL/Rust now uses for
compilation.  This prohibits newer versions of PL/Rust, which might have
a new lint, from executing functions created with an older version
without that lint.

We do, however, continue to require the lints that are set in both the
`$PLRUST_REQUIRED_LINTS` envar and the `plrust.required_lints`
postgresql.conf GUC.

This is a way for the administrator to flat out **require** that certain
lints be applied to **every** function before it can be loaded/executed.

PL/Rust's default stance, however, is that if it was good before, it's
good now, even if new lints are applied going forward.