In response to issue #325, rework the dependency allow-list. (#332)
* In response to issue #325, rework the dependency allow-list.

Generally, the changes here allow a single dependency in the allow-list to have one **or more** versions specified, either as exact (`=`) or bounded ranges (`>=,<=`) or wildcards (`*`).

PL/Rust will pick the largest (newest, hopefully) entry from the allow-list that matches the version the user's requested version.  Depending on how either is written, the allow-list version could be used or the user's requested version could be used.

See the new documentation for more details.